+++ controllers/mover/rsynctls/rsync_test.go
@@ -1163,6 +1163,30 @@ var _ = Describe("RsyncTLS as a source", func() {
 					Expect(job.Spec.Template.Spec.ServiceAccountName).To(Equal(sa.Name))
 				})
 
+				When("The ReplicationSource CR name is very long", func() {
+					BeforeEach(func() {
+						rs.Name = "very-long-name-will-cause-job-name-to-be-evenlongerthan63chars"
+					})
+
+					It("The job name should be shortened appropriately (should handle long CR names)", func() {
+						j, e := mover.ensureJob(ctx, sPVC, sa, tlsKeySecret.GetName()) // Using sPVC as dataPVC (i.e. direct)
+						Expect(e).NotTo(HaveOccurred())
+						Expect(j).To(BeNil()) // hasn't completed
+
+						jobs := &batchv1.JobList{}
+						Expect(k8sClient.List(ctx, jobs, client.InNamespace(rs.Namespace))).To(Succeed())
+						Expect(len(jobs.Items)).To(Equal(1))
+						moverJob := jobs.Items[0]
+
+						// Reload the replicationsource to see that it got updated
+						Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(rs), rs)).To(Succeed())
+
+						Expect(moverJob.GetName()).To(ContainSubstring(string(rs.GetUID())))
+						// Make sure our shortened name is actually short enough
+						Expect(len(moverJob.GetName()) > 63).To(BeFalse())
+					})
+				})
+
 				getSPVC := func() *corev1.PersistentVolumeClaim {
 					return sPVC
 				}
@@ -1690,6 +1714,41 @@ var _ = Describe("Rsync as a destination", func() {
 			})
 		})
 
+		Context("Service handled properly when replicationdestination name is very long", func() {
+			BeforeEach(func() {
+				rd.Name = "very-long-name-will-cause-job-name-to-be-evenlongerthan63chars"
+			})
+
+			It("The service name should be shortened appropriately (should handle long CR names)", func() {
+				// create the svc
+				result, err := mover.ensureServiceAndPublishAddress(ctx)
+				Expect(err).To(BeNil())
+
+				if !result {
+					// This means the svc address wasn't populated immediately
+					// Keep reconciling - when service has address populated it should get updated in the rs status)
+					Eventually(func() bool {
+						gotAddr, err := mover.ensureServiceAndPublishAddress(ctx)
+						return err != nil && gotAddr
+					}, maxWait, interval).Should(BeTrue())
+				}
+
+				// Find the service
+				svcs := &corev1.ServiceList{}
+				Expect(k8sClient.List(ctx, svcs, client.InNamespace(rd.Namespace))).To(Succeed())
+				Expect(len(svcs.Items)).To(Equal(1))
+				rdSvc := svcs.Items[0]
+
+				Expect(rdSvc.GetName()).To(ContainSubstring(string(rd.GetUID())))
+				// Make sure our shortened name is actually short enough
+				Expect(len(rdSvc.GetName()) > 63).To(BeFalse())
+
+				// Service name should be saved in the status
+				Expect(rd.Status.RsyncTLS.ServiceName).NotTo(BeNil())
+				Expect(*rd.Status.RsyncTLS.ServiceName).To(Equal(rdSvc.GetName()))
+			})
+		})
+
 		//nolint:dupl
 		Context("Service and address are handled properly", func() {
 			var svc *corev1.Service
@@ -1719,6 +1778,9 @@ var _ = Describe("Rsync as a destination", func() {
 					err = k8sClient.Get(ctx, client.ObjectKeyFromObject(svc), svc)
 					Expect(err).ToNot(HaveOccurred())
 				}
+
+				Expect(rd.Status.RsyncTLS.ServiceName).NotTo(BeNil())
+				Expect(*rd.Status.RsyncTLS.ServiceName).To(Equal(svc.GetName()))
 			})
 
 			When("spec leaves service defaults", func() {