+++ controllers/utils/utils.go
@@ -18,7 +18,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 package utils
 
 import (
-	"context"
+	"context" //nolint:gosec
 	"fmt"
 	"os"
 	"sort"
@@ -42,6 +42,12 @@ import (
 // Define the error messages to be returned by VolSync.
 const (
 	ErrUnableToSetControllerRef = "unable to set controller reference"
+
+	VolSyncPrefix = "volsync-" // TODO: remove old VolSyncPrefix from mover package?
+
+	JobNameMaxLength     = 63
+	ServiceNameMaxLength = 63
+	LabelValueMaxLength  = 63
 )
 
 // Check if error is due to the CRD not being present (API kind/group not available)
@@ -254,4 +260,33 @@ func UpdatePodTemplateSpecFromMoverConfig(podTemplateSpec *corev1.PodTemplateSpe
 	for label, value := range moverConfig.MoverPodLabels {
 		podTemplateSpec.Labels[label] = value
 	}
+}
+
+// Will return a name with prefix + owner.Name unless it's too long, in which
+// case we will return prefix + owner.UID
+// (This assumes namePrefix + UID is shorter than 63 chars)
+func GetJobName(namePrefix string, owner client.Object) string {
+	return getShortenedResourceName(namePrefix, owner, JobNameMaxLength)
+}
+
+// Will return a name with prefix + owner.Name unless it's too long, in which
+// case we will return prefix + owner.UID
+// (This assumes namePrefix + UID is shorter than 63 chars)
+func GetServiceName(namePrefix string, owner client.Object) string {
+	return getShortenedResourceName(namePrefix, owner, ServiceNameMaxLength)
+}
+
+func GetOwnerNameLabelValue(namePrefix string, owner client.Object) string {
+	return getShortenedResourceName(namePrefix, owner, LabelValueMaxLength)
+}
+
+func getShortenedResourceName(namePrefix string, owner client.Object, maxLength int) string {
+	name := namePrefix + owner.GetName()
+
+	if len(name) > maxLength {
+		return namePrefix + string(owner.GetUID())
+	}
+
+	// No need to shorten, use original name
+	return name
 }